refactor: verify exists IDs

+++ src/main/java/elp/edu/pe/horario/application/usecase/seccion/EliminarSeccionUseCase.java
@@ -1,10 +1,14 @@
 package elp.edu.pe.horario.application.usecase.seccion;
 
 import elp.edu.pe.horario.domain.model.Seccion;
+import elp.edu.pe.horario.domain.repository.CursoSeccionRepository;
 import elp.edu.pe.horario.domain.repository.SeccionRepository;
 import elp.edu.pe.horario.shared.exception.BadRequest;
+import elp.edu.pe.horario.shared.exception.CustomException;
 import elp.edu.pe.horario.shared.exception.DeleteException;
 import elp.edu.pe.horario.shared.exception.NotFoundException;
+import jakarta.transaction.Transactional;
+import lombok.RequiredArgsConstructor;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import org.springframework.stereotype.Service;
@@ -12,15 +16,14 @@ import org.springframework.stereotype.Service;
 import java.util.UUID;
 
 @Service
+@RequiredArgsConstructor
 public class EliminarSeccionUseCase {
 
     private final static Logger log = LoggerFactory.getLogger(EliminarSeccionUseCase.class);
     private final SeccionRepository seccionRepository;
+    private final CursoSeccionRepository cursoSeccionRepository;
 
-    public EliminarSeccionUseCase(SeccionRepository seccionRepository) {
-        this.seccionRepository = seccionRepository;
-    }
-
+    @Transactional
     public void ejecutar(UUID id) {
         try{
             if(id == null) throw new BadRequest("El id no puede ser nulo");
@@ -29,8 +32,12 @@ public class EliminarSeccionUseCase {
                     .findById(id)
                     .orElseThrow(() -> new NotFoundException("Sección no encontrada"));
 
+            // Verificar si la sección está asociada a algún curso
+            if (cursoSeccionRepository.existsBySeccionId(id)) {
+                throw new CustomException("No se puede eliminar la sección porque está asociada a un curso");
+            }
+            // Eliminar la sección
             seccionRepository.deleteById(id);
-
             log.info("Sección eliminada: {}", seccion);
         } catch (Exception e) {
             log.error("Error al eliminar la sección: {}", e.getMessage());
